+++ lib/components/TeethGroup.tsx
