+++ lib/components/Tooth.tsx
